+++ app/Http/Controllers/DriverController.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\Driver\StorDriverRequest;
+use App\Http\Requests\Driver\StoreDriverRequest;
 use App\Http\Requests\Driver\UpdateDriverRequest;
+use App\Models\Bus;
 use App\Models\Driver;
+use App\Models\User;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Request;
+use Symfony\Component\HttpFoundation\Response;
 
 class DriverController extends Controller
 {
@@ -15,19 +17,40 @@ class DriverController extends Controller
      */
     public function index(): JsonResponse
     {
-        $drivers = Driver::all();
-        return $this->getJsonResponse($drivers, "Drivers Fetched Successfully");
+        /**
+         * @var User $user;
+         */
+        $user = auth()->user();
+        if($user->can('View Drivers')){
+            $drivers = Driver::all();
+            return $this->getJsonResponse($drivers, "Drivers Fetched Successfully");
+        }else{
+            abort(Response::HTTP_FORBIDDEN);
+        }
+
     }
 
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(StorDriverRequest $request): JsonResponse
+    public function store(StoreDriverRequest $request): JsonResponse
     {
-        $data = $request->validated();
-        $driver = Driver::query()->create($data);
-        return $this->getJsonResponse($driver, "Driver Created Successfully");
+
+        /**
+         * @var Driver $driver;
+         * @var Bus $bus;
+         * @var User $user;
+         */
+        $user = auth()->user();
+        if($user->can('Add Driver')){
+            $data = $request->validated();
+            $driver = Driver::query()->create($data);
+            $driver->buses()->attach($data['bus_ids']);
+            return $this->getJsonResponse($driver, "Driver Created Successfully");
+        }else{
+            abort(Response::HTTP_FORBIDDEN);
+        }
     }
 
     /**
@@ -43,9 +66,22 @@ class DriverController extends Controller
      */
     public function update(UpdateDriverRequest $request, Driver $driver): JsonResponse
     {
-        $data = $request->validated();
-        $driver->update($data);
-        return $this->getJsonResponse($driver, "Driver Updated Successfully");
+        /**
+         * @var Driver $driver;
+         * @var Bus $bus;
+         * @var User $user;
+         */
+        $user = auth()->user();
+        if($user->can('Update Driver')){
+            $data = $request->validated();
+            if(isset($data['bus_ids'])){
+                $driver->buses()->sync($data['bus_ids']);
+            }
+            $driver->update($data);
+            return $this->getJsonResponse($driver, "Driver Updated Successfully");
+        }else{
+            abort(Response::HTTP_FORBIDDEN);
+        }
 
     }
 
@@ -54,7 +90,16 @@ class DriverController extends Controller
      */
     public function destroy(Driver $driver): JsonResponse
     {
-        $driver->delete();
-        return $this->getJsonResponse([], "Driver Deleted Successfully");
+        /**
+         * @var User $user;
+         */
+        $user = auth()->user();
+        if($user->can('Delete Driver')){
+            $driver->delete();
+            return $this->getJsonResponse([], "Driver Deleted Successfully");
+
+        }else{
+            abort(Response::HTTP_FORBIDDEN);
+        }
     }
 }
